Adding an iOS Appiem driver for the UITests
Very much a work in progress

Much work needs to be done to make the app UI items searchable.
Also Appium does not support invoking selectors in the running app
so many tests will fail for that reason. We will need to find new ways to perform
these actions and queries.